Create main.c
This is the c file of the program. I have not touched this program in a while but it got the job done. I wanted to learn a little about how C handled things and I also wanted to try to make a naive implementation of a genetic program. This code works in Visual Studio 2019 but I don't see any reason why you could not get it to work with any C compiler that is standards compliant. I am not really looking for any contributions to the code and I give a warning that this code does use a GOTO... I think GOTO is fine for use, especially in certain situations.